fit_indices {regsem}	R Documentation

Calculates the fit indices

Description

Calculates the fit indices

Usage

fit_indices(model, CV = F, CovMat = NULL, data = NULL,
  n.obs = NULL)

Arguments

model
regsem model object.

CV
cross-validation. Note that this requires splitting the dataset into a training and test set prior to running the model. The model should be run on the training set, with the test set held out and then passed to CovMat=.

CovMat
If CV=T then test covariance matrix must be supplied. Note That this should be done before running the lavaan model and should not overlap with the data or covariance matrix used to run the model.

data
supply the dataset?

n.obs
Number of observations in the test set for CV.

Examples

## Not run: 
fit_indices()

## End(Not run)

[Package regsem version 1.3.9 Index]