parse_parameters {regsem}	R Documentation

Takes either a vector of parameter ids or a vector of named parameters and returns a vector of parameter ids

Description

Takes either a vector of parameter ids or a vector of named parameters and returns a vector of parameter ids

Usage

parse_parameters(x, model)

Arguments

x
Parameter labels

model
Lavaan model

Value

NULL if undefined input. Else vector of parameter ids

[Package regsem version 1.3.9 Index]