rcpp_grad_ram {regsem}	R Documentation

Calculates the gradient vector based on Von Oertzen \& Brick, 2014

Description

Calculates the gradient vector based on Von Oertzen \& Brick, 2014

Usage

rcpp_grad_ram(par, ImpCov, SampCov, Areg, Sreg, A, S, F, lambda, type2,
  pen_vec, diff_par)

Arguments

par
vector with parameters.

ImpCov
expected covariance matrix.

SampCov
Sample covariance matrix.

Areg
A matrix with current parameter estimates.

Sreg
S matrix with current parameter estimates.

A
A matrix with parameter labels.

S
S matrix with parameter labels.

F
F matrix.

lambda
penalty value.

type2
penalty type.

pen_vec
parameter indicators to be penalized.

diff_par
parameter values to take deviations from.

[Package regsem version 1.3.9 Index]